3.0.2 bugfixing int not working this time

+++ variadic_functions/variadic_functions.h
@@ -1,7 +1,7 @@
 #ifndef HEADER
 #define HEADER
 
-#define getvalu(X)  ((X) > (0) ? (printf(toprint, va_arg(ap, int))) : printf(toprint, (va_arg(ap, int))))
+#define getvalu(X)  ((X) > (0) ? (printf(toprint, va_arg(ap, int))) : printf(toprint, (va_arg(ap, double))))
 int sum_them_all(const unsigned int n, ...);
 void print_numbers(const char *separator, const unsigned int n, ...);
 void print_strings(const char *separator, const unsigned int n, ...);
